add warning window for bids export directory deletion. Update handling for HED for events.tsv

+++ bids_export.m
@@ -278,8 +278,20 @@ end
 % deleting folder
 fprintf('Exporting data to %s...\n', opt.targetdir);
 if exist(opt.targetdir,'dir')
-    disp('Deleting folder...')
-    rmdir(opt.targetdir, 's');
+    uilist = { ...
+        { 'Style', 'text', 'string', 'Output directory exists and all current files will be deleted if continue', 'fontweight', 'bold'  }, ...
+        { 'Style', 'text', 'string', 'Would you want to proceed?'}, ...
+        };
+    geometry = { [1] [1]};
+    geomvert =   [1  1 ];
+    [results,userdata,isOk,restag] = inputgui( 'geometry', geometry, 'geomvert', geomvert, 'uilist', uilist, 'title', 'Warning');
+    if isempty(isOk) 
+        disp('BIDS export cancelled...')
+        return 
+    else
+        disp('Deleting folder...')
+        rmdir(opt.targetdir, 's');
+    end
 end
 
 disp('Creating sub-directories...')
@@ -768,7 +780,7 @@ if ~isempty(EEG.event)
         else                                opt.eInfo(end+1,:) = { 'value'         'type' }; end
         if isfield(EEG.event, 'response_time'), opt.eInfo(end+1,:) = { 'response_time' 'response_time' }; end
         if isfield(EEG.event, 'stim_file'),     opt.eInfo(end+1,:) = { 'stim_file'     'stim_file' }; end
-        if isfield(EEG.event, 'usertags'),      opt.eInfo(end+1,:) = { 'HED'           'usertags' }; end
+        if isfield(EEG.event, 'HED'),      opt.eInfo(end+1,:) = { 'HED'           'HED' }; end
     else
         bids_fields = opt.eInfo(:,1);
         if ~any(strcmp(bids_fields,'onset'))
@@ -794,7 +806,7 @@ if ~isempty(EEG.event)
     end
     
     % reorder fields so it matches BIDS
-    fieldOrder = { 'onset' 'duration' 'sample' 'trial_type' 'response_time' 'stim_file' 'value' 'HED' };
+    fieldOrder = { 'onset' 'duration' 'sample' 'trial_type' 'response_time' 'stim_file' 'value'}; % 'HED' }; % remove HED from default column in events.tsv as HED tags should be put in events.json instead
     newOrder = [];
     for iField = 1:length(fieldOrder)
         ind = strmatch(fieldOrder{iField}, opt.eInfo(:,1)', 'exact');
@@ -937,21 +949,21 @@ if ~isempty(EEG.event)
                         end
                         str{end+1} = eventValue;
                         
-                    case 'HED'
-                        hed = 'n/a';
-                        if isfield(EEG.event, tmpField) && ~isempty(EEG.event(iEvent).(tmpField))
-                            hed = EEG.event(iEvent).(tmpField);
-                        else
-                            if isfield(EEG.event, 'usertags') && ~isempty(EEG.event(iEvent).usertags)
-                                hed = EEG.event(iEvent).usertags;
-                                if isfield(EEG.event, 'hedtags') && ~isempty(EEG.event(iEvent).hedtags)
-                                    hed = [hed ',' EEG.event(iEvent).hedtags];
-                                end
-                            elseif isfield(EEG.event, 'hedtags') && ~isempty(EEG.event(iEvent).hedtags)
-                                hed = EEG.event(iEvent).hedtags;
-                            end
-                        end
-                        str{end+1} = hed;
+%                     case 'HED'
+%                         hed = 'n/a';
+%                         if isfield(EEG.event, tmpField) && ~isempty(EEG.event(iEvent).(tmpField))
+%                             hed = EEG.event(iEvent).(tmpField);
+%                         else
+%                             if isfield(EEG.event, 'HED') && ~isempty(EEG.event(iEvent).usertags)
+%                                 hed = EEG.event(iEvent).usertags;
+%                                 if isfield(EEG.event, 'hedtags') && ~isempty(EEG.event(iEvent).hedtags)
+%                                     hed = [hed ',' EEG.event(iEvent).hedtags];
+%                                 end
+%                             elseif isfield(EEG.event, 'hedtags') && ~isempty(EEG.event(iEvent).hedtags)
+%                                 hed = EEG.event(iEvent).hedtags;
+%                             end
+%                         end
+%                         str{end+1} = hed;
                         
                     otherwise
                         if isfield(EEG.event, opt.eInfo{iField,2})
